# openapi.yaml
openapi: 3.0.3
info:
  title: Resort Backend API (compat + core)
  version: 1.0.0
  description: "Machine-readable contract for the Resort Backend used by the frontend. Includes compat `/api` endpoints, admin/internal headers, and upload specs."
servers:
  - url: "{API_BASE}"
    description: "Runtime server (use VITE_API_URL or REACT_APP_API_URL as API_BASE)"
    variables:
      API_BASE:
        default: "http://localhost:8000"
components:
  securitySchemes:
    AdminKey:
      type: apiKey
      in: header
      name: X-Admin-Key
    InternalKey:
      type: apiKey
      in: header
      name: X-Internal-Key
...
paths:
...
  /accommodations/:
    get:
      summary: List accommodations
      responses:
        '200':
          description: List
          content:
            application/json: {}
    post:
      summary: Create accommodation
      requestBody:
        required: true
        content:
          application/json: {}
      responses:
        '201': { description: Created }
  /accommodations/{accommodation_id}:
    get:
      parameters:
        - name: accommodation_id
          in: path
          required: true
          schema: { type: string }
      responses:
        '200': { description: OK }
    put:
      parameters:
        - name: accommodation_id
          in: path
          required: true
          schema: { type: string }
      requestBody: { content: { 'application/json': {} } }
      responses:
        '200': { description: Updated }
    delete:
      parameters:
        - name: accommodation_id
          in: path
          required: true
          schema: { type: string }
      responses:
        '204': { description: Deleted }
...
security: []